requirements.txt
fastapi
uvicorn
pydantic
chromadb
langchain
langchain-openai
langchain-huggingface
PyPDF2
camelot-py[cv]
requests
python-dotenv
python-multipart
boto3
einops